+++ R/beta_div_test.R
@@ -1518,6 +1518,9 @@ var_par_pq <-
 #'     nperm = 9,
 #'     dbrda_computation = TRUE
 #'   )
+#' 
+#'   plot_var_part_pq(res_var_9)
+#' 
 #'   res_var_2 <- var_par_rarperm_pq(
 #'     data_fungi_woNA,
 #'     list_component = list(
@@ -1527,6 +1530,8 @@ var_par_pq <-
 #'     nperm = 2,
 #'     dbrda_computation = TRUE
 #'   )
+#' 
+#'    plot_var_part_pq(res_var_2)
 #' }
 #' }
 #' @details